fix focusing issues on story content that loads slowly
