+++ clients/android/NewsBlur/res/layout/fragment_readingitem.xml
@@ -5,7 +5,7 @@
 
     <!-- The two views in this top FrameLayout are swapped when HTML5 webview custom content wants to be displayed -->
 
-    <com.newsblur.view.ObservableScrollView
+    <com.newsblur.view.ReadingScrollView
         android:id="@+id/reading_scrollview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -42,14 +42,16 @@
                 android:layout_marginLeft="10dp"
                 android:layout_marginRight="10dp"
                 android:layout_marginTop="12dp"
-                android:scrollbars="none" />
+                android:scrollbars="none"
+                android:focusable="true"
+            />
 
            <include
                layout="@layout/include_reading_item_comment" />
             
         </LinearLayout>
 
-    </com.newsblur.view.ObservableScrollView>
+    </com.newsblur.view.ReadingScrollView>
 
     <FrameLayout
         android:id="@+id/custom_view_container"
